+++ src/app/Profile/requests/[id]/inspection-report/page.tsx
@@ -16,7 +16,6 @@ export default function InspectionReport() {
         
         instance.get(ApiHelper.get("GetUserOrderDetail") + "?OrderId=" + params.id).then((res:any) => {
             setOrderDetail(res);
-            debugger
         })
 
     }
@@ -26,12 +25,17 @@ export default function InspectionReport() {
             const logOut = () => {
     instance.post(ApiHelper.get("Logout"))
       .then((res: any) => {
-        localStorage.clear();
+        if (typeof window !== 'undefined') {
+          localStorage.clear();
+        }
         router.push("/");
-       
       })
       .catch((err: any) => {
-        console.error("Error fetching data:", err);
+        console.error("Error logging out:", err);
+        if (typeof window !== 'undefined') {
+          localStorage.clear();
+        }
+        router.push("/");
       });
   }
     return (
