+++ src/app/Profile/requests/page.tsx
@@ -29,12 +29,17 @@ export default function Profile() {
     const logOut = () => {
     instance.post(ApiHelper.get("Logout"))
       .then((res: any) => {
-        localStorage.clear();
+        if (typeof window !== 'undefined') {
+          localStorage.clear();
+        }
         router.push("/");
-       
       })
       .catch((err: any) => {
-        console.error("Error fetching data:", err);
+        console.error("Error logging out:", err);
+        if (typeof window !== 'undefined') {
+          localStorage.clear();
+        }
+        router.push("/");
       });
   }
